pyproject.toml
[project]
name = "surdo-perevodchik"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.4.1",
    "protobuf>=6.33.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rootutils",
    "sacrebleu>=2.5.1",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "tqdm>=4.66.0",
    "transformers>=4.57.1",
    "typer>=0.9.0",
]

[dependency-groups]
dev = [
    "ipykernel",
    "mypy",
    "pre-commit",
    "ipywidgets",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
testpaths = ["tests"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",    # line too long, handled by formatter
    "B008",    # do not perform function calls in argument defaults
    "C901",    # too complex
    "D1",      # Allow missing docstrings
    "D417",    # Don't require descriptions for all arguments
    "F401",    # unused imports
]
[tool.mypy]
ignore_missing_imports = true

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
split-on-trailing-comma = false
lines-after-imports = 2

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
multiline-quotes = "double"
inline-quotes = "double"
